requirements.txt
kafka-python==2.0.2
redis
requests
pandas~=2.2.3
openai~=1.50.1
Flask~=3.0.3
